Move Worker task RemoveContact to Contact\Remove
